+++ example/lib/screens/home/bloc/bloc.dart
@@ -1,3 +1,3 @@
 export 'package:bsev_demo/screens/home/bloc/Home_bloc.dart';
+export 'package:bsev_demo/screens/home/bloc/home_communication.dart';
 export 'package:bsev_demo/screens/home/bloc/home_events.dart';
-export 'package:bsev_demo/screens/home/bloc/home_streams.dart';