+++ example/lib/screens/second/bloc/second_communication.dart
@@ -1,0 +1,12 @@
+import 'package:bsev/bsev.dart';
+import 'package:bsev/communication_base.dart';
+
+class SecondCommunication extends CommunicationBase {
+  final count = BehaviorSubjectCreate<int>();
+
+  @override
+  void dispose() {
+    count.close();
+    super.dispose();
+  }
+}