+++ lib/bloc_communication.dart
@@ -1,8 +1,0 @@
-import 'package:bsev/bsev.dart';
-
-class BlocCommunication<S> {
-  final void Function(EventsBase) dispatcher;
-  final S streams;
-
-  BlocCommunication(this.dispatcher, this.streams);
-}