+++ lib/bsev.dart
@@ -1,13 +1,12 @@
 library bsev;
 
 export 'package:bsev/bloc_base.dart';
-export 'package:bsev/bloc_communication.dart';
-export 'package:bsev/bloc_view.dart';
 export 'package:bsev/bsev_widget.dart';
+export 'package:bsev/communication_base.dart';
 export 'package:bsev/events_base.dart';
 export 'package:bsev/extensions.dart';
 export 'package:bsev/injector.dart';
-export 'package:bsev/stream_base.dart';
 export 'package:bsev/stream_create.dart';
 export 'package:bsev/stream_listener.dart';
+export 'package:bsev/util.dart';
 export 'package:rxdart/rxdart.dart';