+++ lib/dispatcher.dart
@@ -1,6 +1,6 @@
 import 'package:bsev/bloc_base.dart';
+import 'package:bsev/communication_base.dart';
 import 'package:bsev/events_base.dart';
-import 'package:bsev/stream_base.dart';
 import 'package:bsev/stream_create.dart';
 
 abstract class Dispatcher {
@@ -18,7 +18,6 @@ class CommunicationBlocView {
 }
 
 class GlobalBlocDispatcher implements Dispatcher {
-  static const LOG = "(Dispatcher)";
   static final GlobalBlocDispatcher _singleton =
       GlobalBlocDispatcher._internal();
 
@@ -31,19 +30,20 @@ class GlobalBlocDispatcher implements Dispatcher {
   GlobalBlocDispatcher._internal();
 
   @override
-  void dispatchToBlocs<T extends BlocBase<StreamsBase>>(EventsBase event) {
+  void dispatchToBlocs<T extends BlocBase<CommunicationBase>>(
+      EventsBase event) {
     _blocs.where((blocs) => blocs is T).forEach((bloc) {
       bloc.eventReceiver(event);
     });
   }
 
   @override
-  void registerBloc(BlocBase<StreamsBase> bloc) {
+  void registerBloc(BlocBase<CommunicationBase> bloc) {
     _blocs.add(bloc);
   }
 
   @override
-  void unRegisterBloc(BlocBase<StreamsBase> bloc) {
+  void unRegisterBloc(BlocBase<CommunicationBase> bloc) {
     _blocs.remove(bloc);
   }
 }