+++ lib/injector.dart
@@ -5,12 +5,20 @@ final getIt = GetIt.instance;
 
 typedef T BlocBuilder<T extends BlocBase>(GetIt injector);
 typedef T DependencyBuilder<T>(GetIt injector);
-typedef T BlocStreamBuilder<T extends StreamsBase>();
+typedef T CommunicationBuilder<T extends CommunicationBase>();
 
-registerBloc<T extends BlocBase, S extends StreamsBase>(
-    BlocBuilder<T> blocBuilder, BlocStreamBuilder streamBuilder) {
-  getIt.registerFactory<S>(() => streamBuilder());
-  getIt.registerFactory<T>(() => blocBuilder(getIt));
+registerBloc<B extends BlocBase, C extends CommunicationBase>(
+    BlocBuilder<B> blocBuilder, CommunicationBuilder communicationBuilder) {
+  getIt.registerFactory<C>(() => communicationBuilder());
+  getIt.registerFactory<B>(() => blocBuilder(getIt));
+}
+
+registerSingletonBloc<B extends BlocBase, C extends CommunicationBase>(
+    BlocBuilder<B> blocBuilder, CommunicationBuilder communicationBuilder) {
+  getIt.registerLazySingleton<C>(
+    () => communicationBuilder()..isSingleton = true,
+  );
+  getIt.registerLazySingleton<B>(() => blocBuilder(getIt));
 }
 
 registerDependency<T>(DependencyBuilder<T> builder, {String dependencyName}) {