+++ lib/stream_base.dart
@@ -1,3 +1,0 @@
-abstract class StreamsBase {
-  void dispose();
-}